requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
